file_poller/requirements.txt
watchdog
kafka-python
backoff
pandas
mysql-connector
